Handle exerciseByKey after exercise command (#10490)
* Add regression test for exerciseByKey after exercise

changelog_begin
changelog_end

The discovery of exerciseByKey commands currently doesn't handle the
case where the contract or contract key is modified during the same
transaction. Issuing a regular exercise command in this case is an error
as it is not possible to reference a contract id that is generated
during the same transaction.

* Track contract-keys from exercise events

changelog_begin
changelog_end

Co-authored-by: Andreas Herrmann <andreas.herrmann@tweag.io>