Codeowners: Notify committer maintainers on API changes. (#10557)
* Codeowners: Notify committer maintainers on API changes.

Those of us that work on kvutils and committers based on it probably
need to know when certain packages under "ledger" change.

In addition, we seem to be working on the participant state API more
than anyone else, so I think it makes sense to change governance a
little bit.

CHANGELOG_BEGIN
CHANGELOG_END

* Codeowners: Use team names for directories under /ledger.

* Codeowners: Add the participant team back to the participant state APIs.

* Codeowners: Limit kvutils to the committer team, as before.

* Codeowners: Add kv-participant to the Sandbox project.