participant-integration-api: Fix a flaky test.
The `LedgerConfigurationSubscriptionFromIndexSpec` was flaky due to
stubs not specifying all behavior for `IndexConfigManagementService`.

This fixes the underlying issue by avoiding stubs in favor of fakes,
which means that _all_ behavior must be modelled.

Note: Martin Fowler has an excellent, terse description of [the various
forms of test doubles][TestDouble].

[TestDouble]: https://www.martinfowler.com/bliki/TestDouble.html

CHANGELOG_BEGIN
CHANGELOG_END