sandbox-classic: Reintroduce SqlLedger tests for the mutable index. (#10722)
I previously changed `SqlLedgerSpec` to use the append-only index, not
realizing that we have both test classes. This removes the
(now-redundant) `SqlLedgerSpecAppendOnly`, and copies `SqlLedgerSpec` to
`SqlLedgerOnMutableIndexSpec`. The only differences are:

- the mutable index tests check for synchronous commits, and
- the append-only index tests check for rejection reason details
  (which are not present in the mutable index schema).

Previously I had added some tests to `SqlLedgerSpec` but not
`SqlLedgerSpecAppendOnly`, so they had diverged a little.

I also think this makes it clearer that where the tests diverge, we want
to keep the append-only tests, not the mutable ones.

CHANGELOG_BEGIN
CHANGELOG_END