Prevent concurrent access to recorded request in mock
The code above has been a source of flakes, initially apparently fixed
by the addition of `@volatile` to `submittedRequest` that for some
reappeared on MacOS CI nodes as of recently. This is probably not the
best fix for the issue, as these mocks were probably originally
thought to be used just once and have been adapted to be re-used
across multiple test calls, so feel free to suggest a different
approach.

changelog_begin
changelog_end