Typecheck nested LF type applications more efficiently
This improves performance typechecking in the Scala LF typechecker by
matching on nested foralls & type applications which allows us to
check them with only a single substitution as opposed to the previous
state of doing one for each type application.

This is about a 25% improvement on typechecking performance of
GHC.Classes from 200ms to 150ms.

changelog_begin
changelog_end