Fix bug in MonotonicRecordTimeIT
Reading explicitly from ledger.begin causes the test to fail on a pruned
ledger. The test framework automatically selects a ledger offset
observed at the start of the test anyway, so there's no need to
explicitly put ledger.begin as the starting offset for the checkpoints.

CHANGELOG_BEGIN
CHANGELOG_END