Do backwards looking command deduplication in kvutils.
For pre-execution this is accomplished by storing the previous command max-record-time and comparing to the current command min-record-time in relation to the command deduplication period.

For normal execution we just compare the record times directly, as with the previous behavior.

CHANGELOG_BEGIN
kvutils - Command deduplication respects the deduplication period passed in the submitted command. The command deduplication period represents the minimum ceiling for deduplication and can be extended based on the time skews (min/max)
CHANGELOG_END