[ledger-api-client] - The default command tracking timeout is no longer influenced by the deprecated `deduplication_time` as a deduplication period.
changelog_begin
ledger-api-client - The default command tracking timeout is no longer influenced by the deprecated `deduplication_time` as a deduplication period. Previously if no timeout was being set in the command and `deduplication_time` was set as the deduplication period then the command tracking timeout was the minimum between the deduplication_time and max tracking timeout.
changelog_end