Only run migration tests for stable versions on each PR
The ones that also go through snapshots are still there but only run
on the daily run.

The reason for this PR is that we’ve actually gotten damlc tests
pretty fast now to the point where I’m waiting for the compat tests to
finish on PRs that don’t force conformance test reruns.

changelog_begin
changelog_end