Introducing DbSupport, pulling up DbDispatcher building in the factory hierarchies
This PR prepares the landscape for seamless integration of Index DB with non IndexService related
front components like UserManagementStore.
DbSupport is a simply a DbDispatcher and a StorageBackendFactory. Creation of it comes also in Flyway migrated flavor for capturing common need.
This is a pure refactoring PR: no behavior/feature changes

CHANGELOG_BEGIN
CHANGELOG_END