Refactor docs build
This is in preparation for the split docs build where we publish our
sphinx sources here and then build the final docs combined with the
Canton docs in the assembly repo.

Apologies for the messy diff. The idea here is the following:

1. We bundle together all files used as sphinx inputs in
sphinx-source-tree. This will be what we publish.
2. We factor out the calls to sphinx-build for the pdf and html
build to be as generic as possible and share the code.
3. The existing rules for the html build & pdf docs don’t invoke
sphinx anymore and instead only keep the postprocessing (copying other
html files or calling lualatex).

In a followup PR I’ll also simplify the postprocessing steps and then
finally publish those artifacts as part of our build.

changelog_begin
changelog_end